Management commands cleaned up and made more reliable and informative
